+++ src/examples/compose/index.ts
@@ -1,10 +1,14 @@
 import { run } from '../../core'
+import { styleHandler } from '../../groups/style'
 import { viewHandler } from '../../interfaces/view'
 import { logFns } from '../../utils/log' // DEV ONLY
 import { mergeStates } from '../../utils/reattach' // DEV ONLY
 
 let app = run({
   root: require('./app').default,
+  groups: {
+    style: styleHandler(),
+  },
   interfaces: {
     view: viewHandler('#app'),
   },