+++ src/examples/mori/app.ts
@@ -1,5 +1,5 @@
 import { Context, Component, ev } from '../../core'
-import { StyleGroup } from '../../utils/style'
+import { StyleGroup, clickable } from '../../utils/style'
 import { hashMap } from 'mori'
 import { evolve, get } from '../../utils/mori'
 
@@ -63,12 +63,14 @@ let style: StyleGroup = {
     display: 'flex',
     justifyContent: 'center',
     alignItems: 'center',
+    ...clickable,
   },
   reset: {
     padding: '4px',
     color: 'white',
     fontSize: '18px',
     backgroundColor: '#EA1818',
+    ...clickable,
   },
 }
 let mDef: Component = {