+++ src/examples/simple/app.ts
@@ -1,5 +1,5 @@
 import { Context, Component, ev } from '../../core'
-import { StyleGroup } from '../../utils/style'
+import { StyleGroup, clickable } from '../../utils/style'
 
 import { View } from '../../interfaces/view'
 import h from 'snabbdom/h'
@@ -67,12 +67,14 @@ let style: StyleGroup = {
     display: 'flex',
     justifyContent: 'center',
     alignItems: 'center',
+    ...clickable,
   },
   reset: {
     padding: '4px',
     color: 'white',
     fontSize: '18px',
     backgroundColor: '#EA1818',
+    ...clickable,
   },
 }