+++ src/examples/worker/app.ts
@@ -1,4 +1,5 @@
-import { Component, stateOf, interfaceOf } from '../../core'
+import { Component, interfaceOf } from '../../core'
+import { stateOf } from '../../utils/component'
 import { StyleGroup } from '../../utils/style'
 import { View } from '../../interfaces/view'
 import h from 'snabbdom/h'