+++ src/examples/worker/index.ts
@@ -1,10 +1,14 @@
 import { runWorker } from '../../utils/worker'
+import { styleHandler } from '../../groups/style'
 import { viewHandler } from '../../interfaces/view'
 import { logFns } from '../../utils/log'
 
 // all communicatios are transfered via postMessage
 runWorker({
   worker: new (<any> require('worker-loader!./worker')),
+  groups: {
+    style: styleHandler(),
+  },
   interfaces: {
     view: viewHandler('#app'),
   },